+++ lib/hcloud/resources/server.rb
@@ -92,7 +92,6 @@ module HCloud
     # TODO: use only for creation
     attribute :user_data
 
-    # TODO: use only for creation
     attribute :location, :location
 
     # TODO: use only for creation
@@ -105,8 +104,6 @@ module HCloud
     attribute :networks, array: true, default: -> { [] }
 
     attribute :root_password
-
-    attribute :datacenter, :datacenter
 
     attribute :included_traffic, :integer
     attribute :ingoing_traffic, :integer
@@ -171,7 +168,7 @@ module HCloud
     action :request_console
 
     def creatable_attributes
-      [:name, :automount, :start_after_create, :user_data, :labels, :public_net, datacenter: [:id, :name], image: [:id, :name], location: [:id, :name], server_type: [:id, :name], ssh_keys: [:id, :name], firewalls: :id, networks: :id, volumes: :id]
+      [:name, :automount, :start_after_create, :user_data, :labels, :public_net, image: [:id, :name], location: [:id, :name], server_type: [:id, :name], ssh_keys: [:id, :name], firewalls: :id, networks: :id, volumes: :id]
     end
 
     def updatable_attributes